implement todo sort

+++ app/src/main/java/soen/kgutwice/sana/TodoAdapter.java
@@ -97,7 +97,7 @@ public class TodoAdapter extends BaseAdapter {
     }
 
     // 아이템 데이터 추가를 위한 함수. 개발자가 원하는대로 작성 가능.
-    public void addItem(String todo, String subject, String deadline, String actualCompletedDay, boolean completed, int importance) {
+    public void addItem(String todo, String subject, String deadline, String actualCompletedDay, boolean completed, float importance) {
         TodoItem item = new TodoItem();
 
         item.setTodo(todo);
